backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
openai==1.99.9
python-dotenv==1.1.1